auth-api/requirements/api.txt
alembic==1.7.6
Authlib==0.15.5
email_validator==1.1.3
Flask==2.0.2
Flask-Caching==1.10.1
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-Security==3.0.0
gevent==21.8.0
psycopg2-binary==2.9.1
pydantic==1.9.0
redis==4.1.0
requests==2.26.0
SQLAlchemy==1.4.26
user-agents==2.2.0
python-dotenv==0.19.2
Flask-Limiter==2.1
Authlib==0.15.5
opentelemetry-api==1.10.0
opentelemetry-exporter-jaeger==1.10.0
opentelemetry-exporter-jaeger-proto-grpc==1.10.0
opentelemetry-exporter-jaeger-thrift==1.10.0
opentelemetry-instrumentation==0.29b0
opentelemetry-instrumentation-flask==0.29b0
opentelemetry-instrumentation-requests==0.29b0
opentelemetry-instrumentation-wsgi==0.29b0
opentelemetry-instrumentation-sqlalchemy==0.29b0
opentelemetry-sdk==1.10.0
opentelemetry-semantic-conventions==0.29b0
opentelemetry-util-http==0.29b0
opentelemetry-opentracing-shim==0.29b0